+++ packages/sdk-shared/index.js
@@ -3,5 +3,5 @@ module.exports = {
   testServerInProcess: require('./src/test-server'),
   testServer: require('./src/run-test-server'),
   processCommons: require('./src/process-commons'),
-  ApiAssertions: require('./coverage-tests/util/ApiAssertions')
+  ApiAssertions: require('./coverage-tests/util/ApiAssertions'),
 }